// frontend/react-ts/src/components/common/SensorComponents/types.ts
// src\components\common\SensorComponents\types.ts
export interface SensorData {
  Air_Pollution: number;
  Ambient_Temperature: number;
  Efficiency: number;
  Flow_Rate: number;
  Fuel_Quality: number;
  Humidity: number;
  Load_Factor: number;
  Maintenance_Quality: number;
  Power_Consumption: number;
  Pressure_In: number;
  Pressure_Out: number;
  Startup_Shutdown_Cycles: number;
  Temperature_In: number;
  Temperature_Out: number;
  TimeData: Date;
  Vibration: number;
  id: number;
}









  